+++ cli/src/main/java/org/droneDSL/cli/codeGen/concrete/Task.java
@@ -21,7 +21,7 @@ public abstract class Task {
   public List<Transition> transitions;
 
 
-  public record Point(float x, float y, float z) {
+  public record Point(double x, double y, double z) {
     public String toJson() {
       return String.format("{'lng': %s, 'lat': %s, 'alt': %s}", x, y, z);
     }
